requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
pydantic==1.10.7
numpy==1.23.5 
pandas==2.0.1
Pillow==9.5.0
tensorflow==2.12.0
scikit-learn==1.2.2
opencv-python==4.7.0.72
python-dotenv==1.0.0
sqlalchemy==2.0.15
alembic==1.11.1

